requirements.txt

openai>=1.0.0
chromadb>=0.5.0
streamlit>=1.33.0
python-dotenv>=1.0.1
